% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvalpha.R
\name{generate_mv_data}
\alias{generate_mv_data}
\title{Generate Multi-Valued Data Sets}
\usage{
generate_mv_data(
  type = "nominal",
  n_units = 10,
  n_observers = 3,
  n_labels = 5,
  tpr = 0.8,
  fpr = 0.01,
  card_pmf = c(0.3, 0.4, 0.2, 0.1),
  p_missing = 0
)
}
\arguments{
\item{type}{Data type. One of "nominal", "ordinal", "interval", or "ratio".}

\item{n_units}{Number of units (rows) in data.}

\item{n_observers}{Number of observers (cols) in data.}

\item{n_labels}{Number of possible labels which could be applied to the data.}

\item{tpr}{True Positive Rate. To generate the data, first a latent set of labels. \code{tpr} describes the probability that a latent label will be identified by each observer.}

\item{fpr}{False Positive Rate. To generate the data, first a latent set of labels. \code{fpr} describes the probability that each observer includes each label not in the latent set.}

\item{card_pmf}{Probability mass function describing the cardinality of observed label sets. Length should be less than or equal to \code{n_labels}.}

\item{p_missing}{Proportion of observations that are randomly missing, indicated by \code{NA}. These are distinct from observations which are made, but the observer applied 0 labels, indicated by \code{NULL}.}
}
\value{
A matrix with \code{n_units} rows and \code{n_observers} columns with list elements.
}
\description{
Generate Multi-Valued Data Sets
}
\examples{
generate_mv_data()
}
