% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_generics.R
\name{print.mvalpha}
\alias{print.mvalpha}
\title{Print mvalpha class object}
\usage{
\method{print}{mvalpha}(x, ...)
}
\arguments{
\item{x}{mvalpha object}

\item{...}{additional parameters}
}
\value{
invisibly returns the alpha estimate of an mvalpha object
}
\description{
Print generic
}
