% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{set_ops}
\alias{set_ops}
\title{Efficient Set Operations}
\usage{
set_ops(A, B, type)
}
\arguments{
\item{A, B}{sets (vectors) of elements}

\item{type}{a string describing the data type of the label set. This can be "nominal",
"ordinal", "interval", or "ratio" and is used to select the appropriate distance metric.}
}
\description{
Find the intersection and set difference(s) of two sets all at once and more
efficiently than calling \code{base::intersect()} and \code{base::setdiff()} separately.
Based on this stackoverflow answer \url{https://stackoverflow.com/a/72631719}
}
