% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_table3}
\alias{ex_table3}
\alias{ex_table8a}
\alias{ex_table8b}
\alias{ex_table8c}
\alias{ex_table9a}
\alias{ex_table9b}
\alias{ex_table9c}
\title{Published Examples}
\format{
Each is a multi-valued nominal set with readers as columns and units
as rows.
}
\source{
\url{https://doi.org/10.1080/19312458.2016.1228863}
}
\usage{
ex_table3

ex_table8a

ex_table8b

ex_table8c

ex_table9a

ex_table9b

ex_table9c
}
\description{
These data sets represent the examples found in
and can be used to confirm the
accurate calculation of alpha.
}
\keyword{datasets}
