% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvalpha.R
\name{mvalpha}
\alias{mvalpha}
\title{Estimate Multi-Valued Krippendorff's Alpha}
\usage{
mvalpha(
  data,
  type = "nominal",
  verbose = TRUE,
  n_boot = NULL,
  parallelize = FALSE,
  cluster_size = NULL
)
}
\arguments{
\item{data}{a tibble containing a list column for each coder. Each row represents
an observation unit, and each cell contains a vector of 0 to many values. \code{NA} values
are used to represent missing observations and \code{NULL} values represent the null set
response.}

\item{type}{a string describing the data type of the label set. This can be "nominal",
"ordinal", "interval", or "ratio" and is used to select the appropriate distance metric.}

\item{verbose}{a logical value which toggles whether status updates are printed to
the console.}

\item{n_boot}{an integer representing the number of bootstrapped estimates to calculate
for mvDo. The default, \code{NULL}, will not generate bootstrapped estimates.}

\item{parallelize}{a logical value indicating whether to implement parallelization
using the \code{parallel} package.}

\item{cluster_size}{an integer describing the number of cores to allocate to parallelization.
If \code{NULL} and \code{parallelize = TRUE}, then the maximum number of available cores minus 1 will
be used.}
}
\description{
\code{mvalpha()} calculates Krippendorff's alpha statistic for multi-valued data sets.
}
